+++ app/src/main/java/com/example/projectone/Databases/ProjectDAO.java
@@ -1,7 +1,6 @@
 package com.example.projectone.Databases;
 
 import androidx.room.Dao;
-import androidx.room.Delete;
 import androidx.room.Insert;
 import androidx.room.Query;
 import androidx.room.Update;
@@ -17,6 +16,24 @@ public interface ProjectDAO {
     //for getting all the data
     @Query("SELECT * FROM projecttable")
     List<ProjectTable> selectAll();
+
+    @Query("SELECT A FROM projecttable")
+    List<String> getA();
+    @Query("SELECT VA FROM projecttable")
+    List<String> getVA();
+
+    @Query("SELECT A, Item FROM ProjectTable")
+    List<ProjectTableProjection> getAAndItem();
+
+     class ProjectTableProjection {
+        public String A;
+        public String Item;
+
+        public ProjectTableProjection(String A, String Item) {
+            this.A = A;
+            this.Item = Item;
+        }
+    }
 
     @Update
     void updateData(ProjectTable projectTable);
