+++ app/src/main/java/com/example/projectone/Helper/DatabaseHelper.java
@@ -4,9 +4,8 @@ import android.content.Context;
 import android.os.AsyncTask;
 import android.widget.Toast;
 
-import androidx.room.DatabaseView;
-
 import com.example.projectone.Databases.DatabaseClient;
+import com.example.projectone.Databases.ProjectDAO;
 import com.example.projectone.Databases.ProjectTable;
 import com.example.projectone.Loadschedule;
 
@@ -174,7 +173,6 @@ public class DatabaseHelper {
         ClearTableTask clearTableTask = new ClearTableTask();
         clearTableTask.execute();
     }
-
